feat(billing): 자동결제(billing) 객체를 interface로 정의 (#10)
